queryGroupPage-init
