+++ src/screens/MainIndex.js
@@ -184,6 +184,7 @@ import JobRoleComponent from './Masters/JobRoleMaster/JobRoleComponent';
 import MenuComponent from './Masters/MenuMaster/MenuComponent';
 import DeleteRequisition from './PO/deleteRequisition/DeleteRequisition';
 import MyTicketRefactored from './TicketManagement/MyTicket/MyTicketFilters';
+import QueryGroupMaster from './Masters/QueryGroupMaster';
 
 class MainIndex extends React.Component {
   render() {
@@ -315,6 +316,11 @@ class MainIndex extends React.Component {
               exact
               path={`/${_base}/QueryType`}
               element={<QueryTypeComponent />}
+            />
+            <Route
+              exact
+              path={`/${_base}/QueryGroupMaster`}
+              element={<QueryGroupMaster />}
             />
             <Route
               exact
