+++ src/screens/TicketManagement/custome.css
@@ -1,3 +1,6 @@
 .bgru {
   backgroundcolor: red !important;
 }
+.override-margin-left {
+  margin-left: 0px !important;
+}
